Refactor member types
